Try_catch Implementado

+++ LegacyProject/src/legacy/Usuario.java
@@ -1,14 +1,22 @@
 package legacy;
 
+import java.nio.file.attribute.UserPrincipalNotFoundException;
+import java.security.InvalidParameterException;
 import java.sql.Connection;
 import java.sql.Driver;
 import java.sql.DriverManager;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+import java.sql.SQLException;
 import java.sql.Statement;
 import java.util.Vector;
 
+import javax.management.InvalidAttributeValueException;
+
 import org.apache.derby.jdbc.EmbeddedDriver;
+
+import Exceptions.ContraseñaInvalidaException;
+import Exceptions.UsuarioNoExisteException;
 
 public class Usuario {
 	
@@ -69,30 +77,36 @@ public class Usuario {
 		return res;
 	}
 
-	// public int delete() throws Exception{
-	// 	Driver derbyEmbeddedDriver = new EmbeddedDriver();
-	// 	DriverManager.registerDriver(derbyEmbeddedDriver);
-	// 	Connection mBD = DriverManager.getConnection(""+BDConstantes.DRIVER+":"+BDConstantes.DBNAME+";create=false", BDConstantes.DBUSER, BDConstantes.DBPASS);
-	// 	PreparedStatement stmt = mBD.prepareStatement("DELETE FROM Usuario WHERE login=" + this.mLogin + " AND pass=" + this.mPassword);
-    // 	int res=stmt.executeUpdate();
-    // 	stmt.close();
-    // 	mBD.close();
-	// 	return res;
-	// }
+	public int delete() throws UsuarioNoExisteException, ContraseñaInvalidaException, SQLException{
+		
 
-	public int delete() throws Exception{
 		Driver derbyEmbeddedDriver = new EmbeddedDriver();
 		DriverManager.registerDriver(derbyEmbeddedDriver);
 		Connection mBD = DriverManager.getConnection(""+BDConstantes.DRIVER+":"+BDConstantes.DBNAME+";create=false", BDConstantes.DBUSER, BDConstantes.DBPASS);
+		 // Verificar si el usuario existe en la base de datos antes de intentar eliminarlo
+		 PreparedStatement checkStmt = mBD.prepareStatement("SELECT login FROM Usuario WHERE login= ?");
+		 checkStmt.setString(1, this.mLogin);
+		 ResultSet checkResult = checkStmt.executeQuery();
+		
+		 if (!checkResult.next()) {
+			throw new UsuarioNoExisteException("Usuario no encontrado");
+		}
+
+
 		PreparedStatement stmt = mBD.prepareStatement("DELETE FROM Usuario WHERE login= ? AND pass= ?");
 		stmt.setString(1, this.mLogin);
 		stmt.setString(2, this.mPassword);
 		int res=stmt.executeUpdate();
-		stmt.close();
-		mBD.close();
+    	stmt.close();
+    	mBD.close();
+
+		// Si res es 0, significa que la contraseña no coincide
+		if (res == 0) {
+			throw new ContraseñaInvalidaException("Contraseña incorrecta para el usuario dado");
+		}
 		return res;
 	}
-	
+
 	public int update () throws Exception{
 		//por ahora no nos ha hecho falta actualizar nada...
 		return 0;
